Use readConfig for security and sensitivity analysis
